requirements.txt
langchain
langchain_openai
python-dateutil